Add last sample when logging due signal above threshold
only logging when the signal is above the given threshold
reduces noise in the logged values but it also means that
a line plot can appear as if the values are 'cutting the
corner' at points where the pressure suddenly changes
due to the missing (dropped) samples in the meantime.
This commit makes sure that the sample from the moment
before will also make it to the log if the reason for
logging is 'signal above threshold'.